pyproject.toml
[project]
name = "recipe-generator"
version = "0.1.0"
description = "A smart recipe generator"
authors = [
    {name = "Shahzaib Ali Khan",email = "shahzaib.ali.k39@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard] (>=0.118.0,<0.119.0)",
    "openai (>=2.0.0,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "google-genai (>=1.39.1,<2.0.0)",
    "uvicorn[standard] (>=0.37.0,<0.38.0)",
]

[tool.poetry]
package-mode = false

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
